Add files via upload
Cleaned up a few classes, made packages more specific. Still want to work on DatabaseController to find cleaner alternatives to some structures and algorithms.

For example, I feel that committing changes on TableView cells to actually setting the changes on the Character and Observed List, and then saving it to the sql table is too decoupled - too many steps where I can forget to add code and cause issues...

Still need to read up and understand StringProperties to see if I should be using those.