Add files via upload
Big changes in terms of code structure. There was too much model logic happening in the controller, so I moved it around, placing it in the models instead.

Created a "master model" to handle all model changes. Listeners in controller should only call methods to get data from model and update the view.

More to be worked on